+++ src/CheckoutSdk/HandlePaymentsAndPayouts/Flow/Entities/CardConfiguration.cs
@@ -1,0 +1,13 @@
+using Checkout.Common;
+using Checkout.HandlePaymentsAndPayouts.Common;
+using Newtonsoft.Json;
+
+using System.Collections.Generic;
+using System.Runtime.Serialization;
+
+namespace Checkout.HandlePaymentsAndPayouts.Flow.Entities
+{
+    public class CardConfiguration : PaymentMethodConfigurationBase
+    {
+    }
+}
