+++ src/CheckoutSdk/HandlePaymentsAndPayouts/Flow/Responses/ApprovedPaymentSubmissionResponse.cs
@@ -1,0 +1,24 @@
+using Checkout.Common;
+using Checkout.HandlePaymentsAndPayouts.Flow.Entities;
+
+namespace Checkout.HandlePaymentsAndPayouts.Flow.Responses
+{
+    public class ApprovedPaymentSubmissionResponse : PaymentSubmissionResponse
+    {
+        /// <summary>
+        /// The payment's status.
+        /// </summary>
+        public override string Status { get; set; } = "Approved";
+
+        /// <summary>
+        /// The Payment Sessions unique identifier (only present in CreateAndSubmitPaymentSession response)
+        /// </summary>
+        public string PaymentSessionId { get; set; }
+
+        /// <summary>
+        /// The secret used by Flow to authenticate payment session requests (only present in CreateAndSubmitPaymentSession response).
+        /// Do not log or store this value.
+        /// </summary>
+        public string PaymentSessionSecret { get; set; }
+    }
+}
