+++ src/CheckoutSdk/HandlePaymentsAndPayouts/Flow/Responses/PaymentSubmissionResponse.cs
@@ -23,60 +23,4 @@ namespace Checkout.HandlePaymentsAndPayouts.Flow.Responses
         /// </summary>
         public PaymentMethod? Type { get; set; }
     }
-
-    public class ApprovedPaymentSubmissionResponse : PaymentSubmissionResponse
-    {
-        /// <summary>
-        /// The payment's status.
-        /// </summary>
-        public override string Status { get; set; } = "Approved";
-
-        /// <summary>
-        /// The Payment Sessions unique identifier (only present in CreateAndSubmitPaymentSession response)
-        /// </summary>
-        public string PaymentSessionId { get; set; }
-
-        /// <summary>
-        /// The secret used by Flow to authenticate payment session requests (only present in CreateAndSubmitPaymentSession response).
-        /// Do not log or store this value.
-        /// </summary>
-        public string PaymentSessionSecret { get; set; }
-    }
-
-    public class DeclinedPaymentSubmissionResponse : PaymentSubmissionResponse
-    {
-        /// <summary>
-        /// The payment's status.
-        /// </summary>
-        public override string Status { get; set; } = "Declined";
-
-        /// <summary>
-        /// The reason for the payment decline.
-        /// </summary>
-        public string DeclineReason { get; set; }
-    }
-
-    public class ActionRequiredPaymentSubmissionResponse : PaymentSubmissionResponse
-    {
-        /// <summary>
-        /// The payment's status.
-        /// </summary>
-        public override string Status { get; set; } = "Action Required";
-
-        /// <summary>
-        /// Instruction for further payment action.
-        /// </summary>
-        public object Action { get; set; }
-
-        /// <summary>
-        /// The Payment Sessions unique identifier (only present in CreateAndSubmitPaymentSession response)
-        /// </summary>
-        public string PaymentSessionId { get; set; }
-
-        /// <summary>
-        /// The secret used by Flow to authenticate payment session requests (only present in CreateAndSubmitPaymentSession response).
-        /// Do not log or store this value.
-        /// </summary>
-        public string PaymentSessionSecret { get; set; }
-    }
 }
